[XLA:GPU] Add support for TF32_TF32_F32 and TF32_TF32_F32_X3 dot algorithm to mlir emitters and TF32_TF32_F32_X3 to cuBLAS
We have the support for BF16_BF16_F32_X3. The cl repeats this work for TF32 case.

For the mlir case the algebraic simplifier was touched.
Now it could rewrite dot to multiply+reduce if one of the TF32 algorithms was specified.

Triton already has the support for both TF32 algorithms versions. So, only the dot_algorithm_rewriter was touched for supporting the fallback to cuBLAS.

PiperOrigin-RevId: 700636896